refactor to use common code
