+++ quality_of_service_demo/rclcpp/CMakeLists.txt
@@ -21,7 +21,10 @@ include_directories(include)
 ament_export_include_directories(include)
 
 function(qos_demo_executable target)
-  add_executable(${target} src/${target}.cpp src/common_nodes.cpp)
+  add_executable(${target}
+    src/${target}.cpp
+    src/common_nodes.cpp
+    src/utils.cpp)
   ament_target_dependencies(${target}
     "rclcpp"
     "rcutils"
@@ -33,6 +36,8 @@ endfunction()
 qos_demo_executable(liveliness)
 qos_demo_executable(lifespan)
 qos_demo_executable(deadline)
+qos_demo_executable(interactive_publisher)
+qos_demo_executable(interactive_subscriber)
 
 if(BUILD_TESTING)
   find_package(ament_lint_auto REQUIRED)
