--- quality_of_service_interactive_demo/subscriber_with_qos/package.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format2.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="2">
-  <name>subscriber_with_qos_demo</name>
-  <version>0.7.3</version>
-  <description>Subscriber with configurable QoS policy settings</description>
-  <maintainer email="miaofei@amazon.com">Miaofei Mei</maintainer>
-  <license>Apache License 2.0</license>
-  <author>Miaofei Mei</author>
-
-  <buildtool_depend>ament_cmake</buildtool_depend>
-
-  <build_depend>rclcpp</build_depend>
-  <build_depend>std_msgs</build_depend>
-
-  <exec_depend>rclcpp</exec_depend>
-  <exec_depend>std_msgs</exec_depend>
-
-  <export>
-    <build_type>ament_cmake</build_type>
-  </export>
-</package>
